requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
django-health-check
django-redis
redis
python-dotenv
loguru
requests
celery
psycopg2-binary

pylint
pylint-django
pytest-pylint
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-mock==3.5.1
model-bakery==1.2.1
requests-mock
fakeredis
